+++ src/drivers/purgers/BaseCachePurger.php
@@ -9,6 +9,14 @@ use craft\base\SavableComponent;
 
 abstract class BaseCachePurger extends SavableComponent implements CachePurgerInterface
 {
+    // Traits
+    // =========================================================================
+
+    use CachePurgerTrait;
+
+    // Public Methods
+    // =========================================================================
+
     /**
      * @inheritdoc
      */