+++ src/drivers/purgers/DummyPurger.php
@@ -21,6 +21,14 @@ class DummyPurger extends BaseCachePurger
         return Craft::t('blitz', 'None');
     }
 
+    // Properties
+    // =========================================================================
+
+    /**
+     * @var bool
+     */
+    public $isDummy = true;
+
     // Public Methods
     // =========================================================================