+++ src/services/OutputCacheService.php
@@ -33,8 +33,9 @@ class OutputCacheService extends Component
         // Add cache tag header if set
         $tags = Blitz::$plugin->cacheTags->getSiteUriTags($siteUri);
 
-        if (!empty($tags)) {
-            header('Cache-Tag: '.Blitz::$plugin->settings->cacheControlHeader);
+        if (!empty($tags) && !Blitz::$plugin->cachePurger->isDummy && Blitz::$plugin->cachePurger->tagHeaderName) {
+            $value = implode(Blitz::$plugin->cachePurger->tagHeaderDelimiter, $tags);
+            header(Blitz::$plugin->cachePurger->tagHeaderName.': '.$value);
         }
 
         $value = Blitz::$plugin->cacheStorage->get($siteUri);